requirements.txt
pyspeedtest==1.2.7
speedtest-cli==2.1.3
